# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Sera",
        "description": "Laravel Sera Test Code",
        "contact": {
            "email": "samuelbenedictx@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/customer": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get list of Customer",
                "operationId": "d0dd6e28f3d205b922f4c74e0c96ed5d",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "string",
                                                        "example": "633f78c918ff2f00d70dab58"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example@example.com"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-10-07T00:54:33.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-10-07T00:54:33.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Add Customer",
                "operationId": "c5fb436946edbb648e888fcd496d8031",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "example title",
                                    "email": "example@example.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "_id": "633f78c918ff2f00d70dab58",
                                                "name": "example",
                                                "email": "example@example.com",
                                                "created_at": "2022-10-07T00:54:33.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer//{_id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Detail Customer",
                "operationId": "520e90946e82de3bd44bdadcb13b4a3e",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "_id": "633f78c918ff2f00d70dab58",
                                                "name": "example",
                                                "email": "example@example.com",
                                                "created_at": "2022-10-07T00:54:33.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/{_id}": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update Customer",
                "operationId": "b3c5e181fc954a80c0067e4c12d7aea6",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "example title",
                                    "email": "example@example.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "_id": "633f78c918ff2f00d70dab58",
                                                "name": "example",
                                                "email": "example@example.com",
                                                "created_at": "2022-10-07T00:54:33.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Delete Customer",
                "operationId": "fc59328234596bb3fe63ed940f2f21bc",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer data deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}